Refactor: Extract to JoinWithoutIndexMethods
