+++ lib/rubocop/cop/isucon_cops.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative "isucon/mixin/database_methods"
+require_relative "isucon/mixin/join_without_index_methods"
 require_relative "isucon/mixin/offense_location_methods"
 require_relative "isucon/mixin/mysql2_xquery_methods"
 require_relative "isucon/mixin/select_asterisk_methods"
